alarm_v2/python/requirements.txt
face_recognition>=1.3.0
opencv-python>=4.8.0
numpy>=1.24.0
pyserial>=3.5
